Makefile: Do not set user ID for container processes when using Podman
With Podman the container root user is mapped to the user that starts
the container. This means the file permissions are already correct and
using a different user inside the container results in missing file
permissions.
Only set the user ID for Docker, not when Podman is used through a
"docker" wrapper script.